+++ src/components/icon/icons/github.tsx
@@ -1,18 +1,18 @@
-export default function GitHub() {
+export const Github = () => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
       width="20"
       height="20"
       viewBox="0 0 20 20"
-      fill="inherit"
+      fill="currentColor"
     >
       <g clipPath="url(#clip0_1217_4)">
         <path
           fillRule="evenodd"
           clipRule="evenodd"
           d="M10.0002 0.416626C7.62627 0.417858 5.33013 1.25867 3.52241 2.78873C1.71469 4.31878 0.513273 6.4383 0.132974 8.76829C-0.247325 11.0983 0.218293 13.4868 1.44658 15.5068C2.67487 17.5267 4.58572 19.0464 6.83746 19.7941C7.33423 19.8863 7.52135 19.5784 7.52135 19.3166C7.52135 19.0548 7.51141 18.2958 7.5081 17.466C4.72618 18.0669 4.13834 16.292 4.13834 16.292C3.68462 15.1395 3.02888 14.8365 3.02888 14.8365C2.12145 14.2208 3.09678 14.2323 3.09678 14.2323C4.10192 14.3031 4.63015 15.258 4.63015 15.258C5.52103 16.7777 6.96993 16.3381 7.53956 16.0813C7.62898 15.4375 7.88896 14.9995 8.17543 14.7509C5.95321 14.5007 3.61839 13.6478 3.61839 9.83781C3.60461 8.84972 3.97337 7.89417 4.64836 7.16886C4.5457 6.9186 4.20292 5.90765 4.74606 4.53449C4.74606 4.53449 5.58561 4.26776 7.49652 5.55366C9.13558 5.10798 10.8649 5.10798 12.504 5.55366C14.4132 4.26776 15.2511 4.53449 15.2511 4.53449C15.7959 5.90436 15.4531 6.9153 15.3505 7.16886C16.0276 7.89429 16.3971 8.8515 16.3821 9.8411C16.3821 13.6593 14.0423 14.5007 11.8168 14.746C12.1744 15.0555 12.494 15.6598 12.494 16.5884C12.494 17.9188 12.4824 18.989 12.4824 19.3166C12.4824 19.5817 12.6629 19.8912 13.1696 19.7941C15.4216 19.0463 17.3327 17.5264 18.561 15.5061C19.7892 13.4858 20.2546 11.0969 19.8739 8.76667C19.4931 6.43646 18.2911 4.31692 16.4828 2.78713C14.6745 1.25734 12.3778 0.417077 10.0035 0.416626H10.0002Z"
-          fill="inherit"
+          fill="currentColor"
         />
         <path
           d="M6.68986 16.646C6.68986 16.7267 6.59713 16.7959 6.47791 16.7975C6.35868 16.7991 6.26099 16.7333 6.26099 16.6526C6.26099 16.5719 6.35371 16.5028 6.47294 16.5011C6.59216 16.4995 6.68986 16.5637 6.68986 16.646Z"
@@ -45,7 +45,7 @@ export default function GitHub() {
       </g>
       <defs>
         <clipPath id="clip0_1217_4">
-          <rect width="20" height="20" fill="inherit" />
+          <rect width="20" height="20" fill="currentColor" />
         </clipPath>
       </defs>
     </svg>
