+++ src/components/icon/icons/mouse.tsx
@@ -1,0 +1,19 @@
+
+export const Mouse = () => {
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="25"
+            height="24"
+            viewBox="0 0 25 24"
+            fill="none"
+        >
+            <path
+                d="M12.5 7V12M12.5 21C9.18629 21 6.5 18.3137 6.5 15V9C6.5 5.68629 9.18629 3 12.5 3C15.8137 3 18.5 5.68629 18.5 9V15C18.5 18.3137 15.8137 21 12.5 21Z"
+                stroke="currentColor"
+                strokeWidth="2"
+                strokeLinecap="round"
+            />
+        </svg>
+    )
+}
