+++ src/components/icon/icons/twitter.tsx
@@ -10,12 +10,12 @@ export default function Twitter() {
       <g clipPath="url(#clip0_1036_28)">
         <path
           d="M11.9047 8.46958L19.3513 0H17.5873L11.1187 7.35251L5.956 0H0L7.80867 11.1194L0 19.9999H1.764L8.59067 12.2338L14.044 19.9999H20M2.40067 1.30158H5.11067L17.586 18.7623H14.8753"
-          fill="inherit"
+          fill="currentColor"
         />
       </g>
       <defs>
         <clipPath id="clip0_1036_28">
-          <rect width="20" height="20" fill="white" />
+          <rect width="20" height="20" fill="currentColor" />
         </clipPath>
       </defs>
     </svg>
