+++ src/components/icon/index.ts
@@ -1,0 +1,9 @@
+import { Github } from "./icons/github";
+import { Mouse } from "./icons/mouse";
+import Twitter from "./icons/twitter";
+
+export const Icon = {
+    Github,
+    Mouse,
+    Twitter,
+}
