--- src/components/icons/Medium.tsx
@@ -1,31 +1,0 @@
-export default function Medium() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-    >
-      <g clipPath="url(#clip0_1036_48)">
-        <path
-          d="M20 9.85653C20 12.506 19.556 14.6551 19.0081 14.6551C18.4602 14.6551 18.0164 12.5065 18.0164 9.85653C18.0164 7.20652 18.4604 5.05798 19.0081 5.05798C19.5558 5.05798 20 7.20633 20 9.85653Z"
-          fill="inherit"
-        />
-        <path
-          d="M17.4691 9.85646C17.4691 12.8142 16.2064 15.2129 14.6487 15.2129C13.0911 15.2129 11.8284 12.8142 11.8284 9.85646C11.8284 6.8987 13.0909 4.5 14.6485 4.5C16.2062 4.5 17.4689 6.89793 17.4689 9.85646"
-          fill="inherit"
-        />
-        <path
-          d="M11.2812 9.8565C11.2812 12.9988 8.75578 15.5461 5.6407 15.5461C2.52561 15.5461 0 12.9982 0 9.8565C0 6.71481 2.52542 4.16675 5.6407 4.16675C8.75598 4.16675 11.2812 6.71423 11.2812 9.8565Z"
-          fill="inherit"
-        />
-      </g>
-      <defs>
-        <clipPath id="clip0_1036_48">
-          <rect width="20" height="20" fill="inherit" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-}
