--- src/components/work-entry/index.ts
@@ -1,1 +1,0 @@
-export { WorkEntry } from './work-entry'
