+++ src/domain/landing/pages/landing-home/index.ts
@@ -1,0 +1,1 @@
+export { LandingHome } from './landing-home';
