+++ src/domain/landing/pages/landing-home/widgets/about/about.tsx
@@ -1,0 +1,19 @@
+
+export const About = () => {
+
+    return (
+        <div className="flex flex-col lg:flex-row gap-8 px-8 md:gap-16 md:px-16 justify-center items-center">
+            <div className="flex flex-col gap-2 w-full lg:w-1/2 max-w-3xl items-center">
+            <h3 className="text-2xl md:text-4xl font-medium text-accent">
+                ABOUT
+            </h3>
+            <p>
+                “config.json” is a pro&shy;duct of bran&shy;ding
+                cre&shy;a&shy;ti&shy;vi&shy;ty. It’s a sim&shy;ple and
+                me&shy;mo&shy;rable name, yet com&shy;plex at its core. The
+                pie&shy;ces on this port&shy;fo&shy;lio aren’t any different.
+            </p>
+            </div>
+        </div>
+    )
+}
