+++ src/domain/landing/pages/landing-home/widgets/about/index.ts
@@ -1,0 +1,1 @@
+export { About } from './about';
