+++ src/domain/landing/pages/landing-home/widgets/hero/typewriter.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { Typewriter } from "react-simple-typewriter";
+import { Typewriter as SimpleTypewriter } from "react-simple-typewriter";
 
-export default function CustomTypewriter() {
+export const Typewriter = () => {
   return (
     <h1 className="font-semibold text-5xl sm:text-6xl md:text-7xl lg:text-8xl">
-      <Typewriter
+      <SimpleTypewriter
         words={["config"]}
         loop={1}
         typeSpeed={150}
