+++ src/domain/landing/pages/landing-home/widgets/work/index.ts
@@ -1,0 +1,1 @@
+export { Work } from './work';
