+++ src/domain/landing/pages/links/index.ts
@@ -1,0 +1,1 @@
+export { Links } from './links';
